Simplify jsxProperties type declarations
Instead of overriding the jsxProperties field on each new subclass of
NativeObject modify the generic type "JSXProperties<T>" so that it can
derive the properties parameter from `this`.

Methods and readonly properties are stripped - see previous commit.
Allowed children types are now declared directly as part of the type of
jsxProperties. The "jsxChildren" entries in the JSON API definitions are
therefore no longer needed in the schema.

Documentation updates to follow later.

Change-Id: I069a25c6122c9e7f11bea8d3b305e35f70dab6a0