Update TypeScript compiler to version 3.3
This covers tools, snippets and declaration tests. Only a few snippets
needed adjustement, though it is not always clear why. One had to be
converted to TypeScript. Examples may have to be adjusted separately.

Since TypeScript occasionally has breaking changes in minor version
updates we should stick to 3.3 for now. To support a whole range of
TypeScript versions would theoretically require to run the tests with
all of them.

Change-Id: I35567d5569c3e3b89d12f460070c7d9d04bb415c