Support different vertical alignments in StackComposite
By default the stack of a StackComposite is top aligned.

Setting "bottom" on the last child of a stack composite aligns it and
the entire stack to the bottom.

When also setting "top" of the first child to an offset causes the stack
to use the entire vertical space. To do so one of the children is
stretched, usually the first one to have both top and bottom set
explicitly.

Change-Id: I75e8aee423ff434345ffe7bd7f002e96b4f3850f