Support JSX SFC functions as selector
Both the function itself and its name are now valid selectors in any
place that supports them.

Revise the type declarations for Selectors to allow SFC functions and to
make them more flexible and compact. Note that the TypeScript type
system can not always distinguish between using a filter function as
selector and a SFC, resulting in some (acceptable) loosening of type
safety. This indirectly also affect the siblings method.

Fix #1680

Change-Id: I9c947daeb6d39d3b4bfb60e3f2595e26af7ffa18