Turn "$" in to a multi-purpose helper function
It's no longer an alias of WidgetCollection, but it can be used much in
the same way. However, it now can also produce strings like in a
TextView, it acts a shortcut for `tabris.contentView.find()` (useful for
snippets) and as way to obtain NativeWidget objects (useful for
debugging).

There is some undocumented error tolerance when using it as a JSX
element since the TypeScript compiler can not check the type of JSX
children.

Adjust two snippets to demonstrate the new features.

Change-Id: I41cfe16db7063d026802218b589ecdf43d3d9844