Always make JSX functions take children as an attribute
The way the TypeScript compiler handles children of JSX elements
requires us to accept them in two different ways, as an attribute and as
varargs/array. That is confusing when implementing custom JSX functions
or factories, so normalize this behavior to always map the children to a
"children" attribute. This is the preferable over the alternative
because it allows for more type safety.

Adjust all jsx factories accordingly. Also start using JSX "attributes"
and "properties" as distinct concepts, because they are. Further changes
will be required to make this separation complete.

Change-Id: I6ee4fd40769b71b79d056d1136f8df7771612f08