Rename "jsxProperties" to "jsxAttributes"
"Properties" are fields of objects that may be created by JSX syntax,
while JSX "attributes" are the assignments done in a JSX tag. They
largely overlap, but they are not identical. Most significantly the
listeners given via JSX are passed to identically named registration
methods, not set as a properties. To make these differences less
confusing rename "jsxProperties" to "jsxAttributes" and the
"JSXProperties" mapped type to "JSXAttributes".

Change-Id: I802357f026badc2e38fca98277c7a78c2ec9c59a