[DOC] Fix all broken links
As discovered by the check-links.js script in the tabris-js-docs repo.

In some cases the linked content was outdated (e.g. console), so it was
removed instead of fixed.

Do use console.log instead of console.warn in generate-doc.ts to avoid
node printing to stderr since that causes issues with the tabris-js-docs
update.js script.

Change-Id: Id32be71878067393362c035524da100a73bd620f