Make global JsxProcessor instance exchangeable in TypeScript
This will be used in the tabris-decorators project.

The global "JSX" object needs to be declared as a name space in
TypeScript, which can unfortunately not be a class instance
(JsxProcessor). Instead provide a "install" method on the name space
object to configure which processor to use. The symbols used by the
processor become part of the name space and are therefore not
exchangeable, but they don't need to be.

Tweak grunt clean task to fix issues with VS Code preventing the build
directory from being deleted.

Change-Id: Id07b3dd74d850f06faab3aa0ca92fd2bf03d9865