Revise handling of vertical spacing in StackLayout
The previous behavior was adding up the values of neighboring top/bottom
child properties and the stack layout spacing. This is not entirely
consistent with the behavior of the left/right/centerX properties and
occasionally also inconvenient. Now setting top or bottom overrides the
spacing.

Setting conflicting top/bottom values is resolved by choosing the
greater value unless one of them is negative, which results in a spacing
of zero pixels. This is implemented as error tolerance, not as a typical
use case.

Change-Id: I9efa37be14c3796a5b93f59bb6a54be816583846