Add JSX shorthands for layoutData presets
These basically allow omitting the attribute ("layoutData") when setting
it to a preset string in JSX. Technically it is setting an
(non-existing) attribute named after the given value to true, but the
JSX factory of Widget re-interprets it as a shorthand.

Since multiple layoutData values can actually be merged no
contradictions have to be excluded. If the merged object is inconsistent
it will be handled by the existing layouting code. Also ensure the
layoutData property is set before any delegate layout properties so
merging those always works as expected. (This is technically an existing
issue unrelated to JSX.)

The mechanic is built in to JSXProcessor and can be re-used by other JSX
elements in the future. The type declarations are hard-coded in to the
JSXAttributes interface since it would be difficult to integrate it
otherwise without messing up other "jsxAttributes" declarations on
Widget subclasses.

Change-Id: I9002ae6c859c420966ce1173e6495c74d8aab95a