Remove "implicit" bottom alignment stack layout feature
In stack layouts the entire stack could be moved to the bottom by
setting "bottom" to 0 on the last child. The main issue with this is
that it makes using the more important "stretch" feature more
complicated than it needs to be. Now the condition for stretching a
child vertically is "just" to set top and bottom together. Moving
everything to the bottom can still be achieved by adding a placeholder
and stretch it.

Documentation to be updated later.

Change-Id: I58235e3e5bd4f7110276f706296792a66a672a8f