Add TextInput properties "style" and "floatMessage"
The "style" properties allows to set the visual appearance of the
TextInput. The supported values are "outline", "fill", "underline" and
"none". In all but the last style the message hint is able to float
upwards when the user focuses the TextInput. To suppress this behaviour
the "floatMessage" property can be set to false.

With the new styles and in order to be consistent with iOS the
"underline" style does not apply the background property anymore but
only the borderColor. This allows us to remove the redundant
`fillColor` property.

Change-Id: I1819742e5280342773f02063ccaa00e0bfc6c20a