Document Module class and implicit local module variable
Document Module class including the static API used by the CLI for
side-loading.

Also document the variables injected by the module system (e.g. require,
__dirname, etc.). They have to be declared global even though the actual
variables are local to each module. This is "fixed" by putting the root
module variables in the actual global context. Within each module they
are shadowed, but using the interactive CLI they can now be used to
inspect specific modules.

The tsconfig.json of the snippets has to be adjusted to exclude the
installed node typings which would clash with the now global require
function.

Change-Id: I045656e32feea876ca5f272522938d21febf1164