Skip type declarations verification in non-release builds
These take a very long time and in the majority of all commits the type
declarations do not change. In case they do change it is the
responsibility of the committer to run the "quickverify" task. On a
release build the typings are still checked in case someone neglected to
do so. The nightly build also calls quickverify explicitly.

Change-Id: I0a57cf081874a15a57b7f3951c761e8beb201f57