Temporarily cache property values for native get in NativeBridge class
This prevents code like `widget.bounds.top + widget.bounds.height` from
unnecessarily making two native get operations.

This is a fairly high-risk change since it assumes that within one
native-to-js call no native properties may change spontaneously.
However, this seems to be the case for all properties since the UI
thread is blocked during that time.

The cache is attached to the NativeBridge rather than the NativeObject
because it is far easier to cover all scenarios and clear the cache with
little effort.

Change-Id: I132416a673f1e125326b0822579789b7af894685