Cache native event parameter for native get in NativeBridge class
The values for changing properties are given on a change event object,
but may also be obtained from the widget. In this case they are
needlessly requested from the native client again.

Let tabris._notify use the event definitions created via
NativeObject.defineEvent to extract an updated property value from the
event parameter. In the case of bounds this causes some ambiguity on the
resize event definition since the changing value is obtained once from
the raw event parameter object and once from the final event object, but
it works for both cases.

Change-Id: I6e17d37c2b1d703716c8f5b1a6713630a8bf7901