Proxify exported singletons
This fixes a fundamental issue with re-initializing the tabris module in
a unit test (usually node) environment.

If a test calls tabris._init to reset the entire tabris module state it
also creates new singletons (e.g. tabris.contentView). However,
application modules may already have imported these from the tabris
module and would therefore hold outdated references.

To resolve this, export proxy objects with handler that do nothing but
forward any interaction with the proxy to the *current* internally
created singleton. This is a sort of hot-code-replacement approach that
has no effect in production, but allows switching the singletons with
new instances when running tests.

Proxies are created as part of the Tabris class, but the actual
singletons are still created in the main.js _init method. The native
contentView property of the Tabris object is now also set there.

Change-Id: I697c50230b71abb1d496297cf94374f774ea24bf