Fix collectionview-scroll snippets on iOS
The collectionview-scroll snippets use cell.bounds to calculate the
position of related views. On ios the cell.bounds return 0 instead of
the actual value.

See https://github.com/eclipsesource/tabris-js/issues/1868

To work around the issue we use the cell.absoluteBounds which works in
this snippet because the CollectionView is positioned at 0,0 of the
contentView.