Add properties font, style, message and floatMessage on Picker
The `Picker` now allows to show an empty state. This unselected state
can be filled with a `message` text similar to a `TextInput`. The
initial `selectionIndex` is therefore `-1`. It can also be set to `-1`
which shows the empty state or the `message` respectively.

The new Android only properties `style` and `floatMessage` behave
similar to their `TextInput` counterpart as they allow to theme the
`Picker` and to control the floating behavior of the new `message`
property.

In addition we remove the now deprecated property `fillColor` which is
replaced by the standard `background` property (again similar to the
`TextInput`).

Finally we introduce the `font` property which had not been reinstated
when it had been removed from `Widget`.

Change-Id: I8d65f4e900ba395adb9565ef81c039f6ffd3bdb3