Compensate inconsistent column mapping in stack traces
There seems to be a difference in how some source map generators (in
this case TypeScript tsc) designate the begin of a code region after
indention versus the stack traces created at runtime. With this change
columns in stack traces that would be outside of a code region due
indention are still mapped.

Change-Id: I75d481a2037a18023d5f2c616b70ebfe28510e94