Fix Tab appear/disappear events
These events are built internally on the selectionChanged event of
TabFolder, but that event was only fired when there was an explicit
listener registered for it.

To keep the order of events
(select->selectionChanged->appear/disappear), do not use a
selectionChanged listener but opt to explicitly trigger selectionChanged
even if no listener is attached. Also always native-listen to the select
event which the selectionChanged event is based on.

Fix #1873

Change-Id: I86ca2fe59f84e4edf4df3c9f64a76f2f0515d651