Rename .jsx snippets with contentView import issues to .tsx
Snippets with .jsx extension fail with an error when the contentView
is accessed before the traditional .append() call with the message

error TS2339: Property 'append' does not exist on type 'typeof contentView'.

This is likely a tsc bug. Rename the snippets imageview-gallery.jsx and
timer.jsx to .tsx to work around it.