Allow to request app runtime permissions
Permissions serve to protect the privacy of a user. In order the access
features like the devices location or camera, the user has to grant
permissions to access them.

The change introduces the tabris object `tabris.Permission` which is
available from the `tabris` object. The `Permission` object allows to
check and requests permissions.

Closes https://github.com/eclipsesource/tabris-js/issues/1896

Change-Id: I885562d568a2645ed218099d640b22f2659058f3