Explicitly document fetch() via api-json
Extend declaration and markdown generators to be able to handle
interfaces definitions. For the API documentation these are not
rendered as separate files but in-line (possible multiple times) where
they are used as parameters. They can currently not be used as property
values. (No need.)

This change also fixes a number of mistakes in the original fetch
declarations (specifically the request init object and some methods
actually being static) and adds more comprehensive test.ts files.

Change-Id: I66ee09a52152711c73263f485b51d5b6d08ff780