Support automatic snippet inlining in generated API documentation
The doc generator will look for files with the same base name as the
.json file and a file extension of 'js', 'jsx', 'ts', 'tsx' or 'console'
and include them below the description separated by headings. There can
be multiple examples per file by separating them with comments.

Extract all code examples previously placed in the '.md' files of the
same name to script files. These files are now also linted (using
ts-lint), but not compiled. Because of this various problems
had to be fixed in the examples.

Change-Id: Id72ef59be2f44553ec68ee4cf01c320ada877ed9