Revise property type system
This change was necessary to allow properties that return the same
objects that was set. This is important for types that wrap NativeObject
instances (specifically Blob), since there is no registry for them
(unlike NativeObject).

To allow this the new "type" specification now has a "convert" function
in addition to a "encode" and "decode" function. The convert function
can check and/or convert the set value to the one that should be
returned later. Ideally it just passes it through. "set" and "get" are
no longer supported since the existing usage of them don' fit the new
system and make any future changes very hard (like this one).

property-types.js is completely rewritten for more consistency and
better IDE/jsdoc support. The 'choice' and 'nullable' types are removed
in favor of additional configuration keys in
NativeObject.defineProperty(options).

Adjust all usage of the 'set', 'get' and 'encode' options. Adjust all
tests. Add more tests for NativeObject.defineProperty.

Change-Id: I2cd94dd9ffba22b7666c5ddcf55049a5a2ed36d0