Ensure TabFolder children are already known to the client when selected
Normally rendering of the internal "children" property is delayed until
it can't change anymore. However, this can cause the "selection"
property of TabFolder to be set to a Tab that is not yet a child of the
TabFolder from the native clients point of view. By skipping the queue
this can be avoided. It's technically less efficient, but in practice
this should never have any real inpact.

Change-Id: Ifa88d6d649828de512f627293cbb4b40fe6689f3