Fix cell/item-index mapping in CollectionView after remove/insert call
Previously the internal cell/item-index mapping (needed by the
"itemIndex" and "getCellByItemIndex" methods) was only updated in the
"updateCell" event. However, a remove or insert operation shifts all
indices greater than the start index of the operation. Since no
"updateCell" event is fired by the native client for the cells currently
associated with these indices they need to be shifted explicitly on the
JS side.

Change-Id: I4d2083eb03c968c1cb6fd5a4a71d45bd43e309db