Fix Picker caching incorrect selectionIndex value
The picker "select" event still has a name translation of the
"selectionIndex" parameter to "index". This confuses the global property
caching logic that normally requires the parameter name to be the same
before and after the event is forwarded to the target NativeObject. The
event definition is now changed so both parameter names are accepted.

Change-Id: Idad8c7bc563149744c7ada09170c83f85f316886