Add documentation for List, ListView/Cell and ItemPicker
These classes are provided by the `tabris-decorators` module, but
documented here to take advantage of the documentation generator. The
generator itself is updated to be able to deal with the JSX syntax used
by ListView and ItemPicker.

Since the TypeScript declarations for these classes are provided in the
`tabris-decorators` module dts-generator is made ignore them via the new
`markdown_only` flag.

Change-Id: Id6ec6f92d1f5cf7a060fc7e949b5a71e8fc8b436