Completely migrate from tslint to eslint
tslint is being deprecated, so lint all code including TypeScript/JSX
with eslint and the appropriate plug-ins.

Various adjustments are required as the result of slightly stricter
linter rules. Most are related to formatting, but also unhandled
promises and shadowed or unused variables.

test/typescript/* is not linted because the setup would be too complex.
$.jsx is not linted because this seems to trigger a bug in the eslint
TypeScript parser.

All linted files need to be included in a tsconfig.json. The root
tsconfig.json is changed accordingly. It is only used by the IDE
otherwise, so changes are not risky.

Change-Id: I3e473dc42cd97e838752ced37c63d7e7382fdb1a