Add tabris.app.share() support
The share api asks the operating system to share the data with another
installed app. The behavior of this API follows the [W3C Web Share API]
(https://www.w3.org/TR/web-share/). It supports to share text, url and
files as well as to provide a title to show in the share dialog.

Change-Id: I53b589b5977e906461578a8b4fc85ce58a0b3779