Revise Response.blob() and Response.arrayBuffer()
Multiple fixes for this function:

1) Do not make unnecessary copies of the contained arrayBuffer. Since
the buffer can only be returned once, it is not an issue if it is
mutated later on.

2) Do encode text responses. So far both methods assumed that they are
only called when the response type is string. If it isn't this caused
a meaningless exception. Now it converts the text with the default
encoding.

3) Do set file type if available. Unlike ArrayBuffer a Blob can contain
type information. If it the "content-type" header is set it will now be
passed on to the Blob constructor.

Change-Id: Ie9bce6618d95546f9faa447899b545c80e7ff484