Implement class "Resources"
This class will provide the basis for managing theming resources such as
fonts and colors based on JSON. It allows to select, verify, convert,
inherit and cross-reference values based on the platforms name, screen
density and locale. Dynamic criteria such as device orientation can
currently not be supported since the resulting resource dictionary is
immutable.

Change-Id: Id6425f540684b4c0c2bd2a09b728e175ab07d434