Implement class "ResourceBuilder"
ResourceBuilder provides a type-safe way to create resource packs of any
type (with or without conversion) with the absolute minimal amount of
code. No subclassing needed.

The resource configuration options become part of the raw data
dictionary (prefixed with "$" to make them distinct from resource names)
so they can be included in a JSON resources file.

The in-between step of creating a "ResourceBuilder" instance (as opposed
to Resources.build directly creating the Resources instance) is
necessary since the TypeScript compiler can not infer the type of the
Resources and their names in the same call.

Change-Id: Ieabbd5a64d7ea391c3b88cb1a92679bf8f3fc43d