Fix source mapping for absolute paths
As of CLI version 3.4 sideloaded modules are known to the JS engine by
their absolute path on the developer machine. They may can either be
unix-style paths starting with "/" or windows-style paths starting with
a drive letter. This fact is used to detect the expected format so that
the path util methods can then be called with an option to indicate how
to treat the url-string.

Change-Id: Id68ddbc7f2fa999b8eb72f3f4a828e1d0cb98cbd