Unify logging of rejection reasons
Previously, rejections returned in event callbacks would break logging
if their results didn't have a toString() function.

Extract the formatting approach for rejections used in Promise to an
util function and use it for event rejections as well. It can handle all
kinds of reasons and ensures that stack traces will be printed when the
rejection reason is not an error.

Update the Promise rejection reason formatting to reuse code in
util-stacktrace ensuring that the original stack trace will be printed
in case the stack trace filter returns an empty stack trace (e.g. when
the stack trace originates from Tabris.js).

Change-Id: Ib856d5e2811138d6819a62898caeea8b736bc833