Update resource-management.md
This adds a section on how to configure a project with resource
dictionaries ("Setup"), explanations for the existing examples, another
example for image resource dictionaries and various tweaks to existing
texts.

What is still missing is an explanation on how to use resource
dictionaries with dependency injection, which can be added at a later
date.

In one instance the article links to external documentation on module
resolution strategies. Since commonjs.org seems to be completely dead
now the node.js documentation is used instead. The module.md article is
updated accordingly as well.

Minor tweak to Resources.json API documentation.

Fix one dead link in debug.md

Change-Id: I3ac0b98c93e32216e6941b7f92583a0f71fff513