Ignore layout flush when already in a flush
Previously an exception was thrown. However, if the flush was triggered
by something else than a layout change it could be safely ignored.
Instead, print a warning if something is added to the layout queue
during a layout flush and ignore it. This should never happen unless
low-level or internal code is manipulating the widget tree within a
layout flush. If it does happen the worse case scenario would now be
that the changes are only partially applied.

Change-Id: If0ad753efb5231fa62bca21514f3b299692d5a3e