Add public utility function "checkType"
This function is now used internally for parameter type checks were
possible, but also exported for use in applications, especially in
TypeScript compiled projects. One snippet was adjusted to use it.

This is the first of possibly many general-purpose utility function
offered by tabris, so a new JSON doc file was created for it.

Adjust error messages expected in unit tests where required.

Change-Id: I99958766be1b3acc4ab7a40f0c1e8cd0ef6aabf8