Fix proxified service classes not lazy-initializing event interface
This issue was introduced by e0bb314bf3cc8ddb6971fe3dcd1e13ab9a88efbb.
The root cause was that the Object.assign method used in that commit did
not actually copy the given object as expected. The fix is to modify the
original, as it is a safe copy already.

Change-Id: Iabf5a439358f4bcdf3551937d43b3646c2c8b29d