Introduce public util function "Set"
This function helps creating type-safe "Properties" objects when the
target type can not be inferred by the IDE/compiler, specifically when
using the apply method. The apply method can also perform a runtime
check that objects created with "Set" are targeting the selected widget.

The function name starts with an upper case so it can also be used as a
JSX element later on. This requires a minor change to api-schema.json to
allow such method names.

Minor JSDoc fix in Widget.js

Change-Id: I678b5ea343c9d7ca795648c789c2c916fd3b7982