Allow listeners to be registered via "apply" method
This new way to register event listeners uses the same logic and code
as JSX elements. That code is moved to Listeners.js for better sharing.

Listeners registered via JSX/apply are treated like event-attributes in
HTML, i.e. there can only ever be one listener registered this way for
any given type. This behavior is necessary for apply to keep behaving in
a deterministic way. The "Set" method is adjusted accordingly.

This does not affect existing event API.

Do not refer to the object given to apply as "properties" in the
documentation, it is incorrect in several ways. Use "rules" instead.

Also fix "Set" method not working with abstract classes.

Change-Id: I616d14212b771a6beb2ec06e9612fc02ed0d8dcc