Allow passing rules to "apply" as a callback
Instead of giving the ruleset directly to the function it can now be
given as a callback returning the actual object. The callback is invoked
with the widget itself. This allows taking properties of the widget in
to account (such as "data") even when the ruleset is defined elsewhere.

It will also allow to automatically recomputing the rules due to
property changes in the future.

Change-Id: I941b9d4076b81384044fea9af41de5b7b44fa730