Add "apply" option "trigger"
This option makes the apply method store the given ruleset so it can be
applied again at a later time. Depending on the value of "trigger" this
may either be when the host widget fires a specific event, or when the
apply method is called with the name of the trigger.

The documented API for the trigger feature is actually stricter than
the implementation, allowing any string or symbol as trigger. That is a
"hidden" feature to be used in tabris-decorators, and also allows
changing how certain values for "trigger" are handled in the future
without breaking the API contract. In addition it improves TypeScript
support.

The JSON docs now use mostly pure strings to define apply parameter
types since the docs generator was not producing satisfying results
otherwise. This should be a temporary solution.

Change-Id: Idbf667d59af92ae76360f755c43eed70d6accd41