Fix "asFactory" passing original instead of proxy to JsxProcessor
The "asFactory" function creates a proxy that catches calls to 
the constructor that do not use the "new" keyword. It then 
passes on the constructor to the JsxProcessor. However, that
processor did get the original constructor, not the proxy. 

This caused issue in tabris-decorators which manipulates the 
proxy. These changes need to be in effect when the JsxProcessor 
calls the constructor. Therefore the processor needs to get 
the proxy object, not the original.

Now the proxy is given to the JsxProcessor.

Change-Id: I037a69c3aef2515b5e776fceb3a3b3418bbee012