[Doc] Describe new widget factory api and functional components
widget-basics.md is updated to cover factories.

selector.md is updated to cover the enhancements of the "apply" method.

The jsx.md article is reworked to the declarative-ud.md article which
covers both JSX and factories.

Also fix some broken links.

One code snippet in declarative-ui.md has an uncommon formatting
"(buttons={ {ok: 'OK'} }" to prevent double-curly-braces which are
detected as template syntax in the markdown-to-html transformation.

Change-Id: Ic60e49c102030033ff6a10a39894dced07a24cdf