Support TypeScript 3.9 and 4.x
To make tabris TypeScript 4.x-compliant some changes are needed in the
type declarations:

* The "properties" parameter of "set" overrides need to assignable to
their respective type in the super class. This is not the case if
properties are explicitly excluded via "Omit" but included in the super
class. Various other mechanism to exclude read-only properties either
don't work with specific TypeScript versions or cause breaking changes.
Therefore these overrides are removed. Regrettably this means that it is
now possible (at compile-time) to use "set" to set read-only values,
though they are still excluded from the constructors.

* The same issue applies to jsxAttributes, but here a workaround is
possible: Instead of omitting attribute they are joined with "never".
This means they are still auto-suggested by the IDE, but actually using
them causes a compile error. This workaround is possible because it is
assumed that the JSXAttribues<T> type is (unlike Properties<T>) is never
used outside of class declarations.

* NativeObject property "_nativeType" must be declared with an explicit
getter to able to override it with one, as is the established way to
define it in code. To allow this the declarations generator and api json
schema need to be extended. This includes a minor unrelated fix in
api-schema.d.ts

* d.ts files may can longer be linted by eslint. The reason for this is
unclear.

Change-Id: I58fbceab2df94f03ebe658173965d5fdc7b60751