Add a Observable implementation
Based on API described in this standard proposal.
https://github.com/tc39/proposal-observable

Also mostly in line with RxJS behavior.

Type declarations are partially shared with implementation by merging
the file Observable.types.d.ts in to the tabris.d.ts. This is currently
done explicitly in Gruntfile.js, but should be automated at some point.

Observables will be integrated in to other Tabris APIs later on.

Change-Id: I19012d903029cb4f00408073fb6039ffb88ac330