Migrate modules related to event handling to TypeScript
This excludes code related to DOM event handling.

Where "Events" was previously a plain object used as a mixin it is now
a class. This allows using access modifiers. NativeObject now inherits
directly from this new class. For the cases where "Events" needs to be
used as a mixin a copy of the class' prototype is exported.

Fix declaration for global tabris object in internals.d.ts. The
"TabrisMain" type from main.js was not properly recognized.

Referencing "Promise" in TypeScript code seems to have triggered an
old rollup issue[1]. The only way to resolve it seems to be to assign
Promise to a variable.

[1] https://stackoverflow.com/questions/45549689/prevent-rollup-from-renaming-promise-to-promise1

Change-Id: I7700b02eca6fe46b1ee211354efb9b2b2949c0e6