Rename util function "Set" to "Setter"
"Set" is also a built-in type of EcmaScript. This was known and
accepted, but the IDE tooling was suffering more from this than
anticipated. Therefore it is now exported as "Setter". The old name is
still valid for backwards compatibility.

Also migrate the implementation and tests to TypeScript.

Move the documentation in to a separate JSON file.

Change-Id: I8a632ba649feb520bbfb7f0da45298e96c01733e