Make Setter a JSX element
This allows inlining larger value expressions in to a JSX tree.
Whether this approach is better than extracting attribute values is up
to the user. For the framework it is meant to maintain feature-parity
between JSX and factory UI declarations without resorting to ugly
multiline attributes.

This feature requires the type declarations to allow any JSX element to
have children. While this reduces type-safety a bit, it was never
very reliable in this regard anyway.

Tabris.js and JsxProcessor.js are migrated to TypeScript.
JsxProcessor.ts was extended to support Setter, while Tabris.ts was
only migrated to fix a compile error.

eslint rule "typescript-eslint/explicit-member-accessibility" was
relaxed temporarily to allow public getters without "public" keyword.
This should be enforced once this style is used everywhere.

Change-Id: I1b048c229f15c64a472ead8037d79bd33d0325d6