Add attributes to <Apply> element for more flexible usage
<Apply> can now have attributes that allow using it similar to
Setter (as a function). That allows for shorter syntax in some cases.

To make better use of this a composite can now have multiple Apply
elements which are merged in to one. This is achieved by allowing
multiple Setter elements for the same attribute if they are all setting
arrays. This is an undocumented feature for now since it is really only
useful for this one scenario.

Change-Id: Iae39a486d47d99b06fa99c2a55a19b6345ccc680