Don't skip native "dispose" operation for children excluded from layout
When a composite is disposed no individual "dispose" operations are sent
to the native client for the children. This is because the native client
is already aware of the children and can dispose them directly.

However, in case "excludeFromLayout" is set to true the widget is no
longer attached to it's parent on the native side. This caused a
memory leak since these children effectively become orphans, marked
as disposed in JavaScript but still in memory on the native side.

Change-Id: Iaa2eead31251afdd0355a2007fb4b313ebcea805