Fix Popups not firing "close" event when closed programmatically
The native side sends a close event if the popup is closed either by
user input, or when disposed from JS, which is all that the "close"
method does. However, disposed native objects can not emit events, so
this is ignored and the event is missing.

Refactor Popup (migrated to TypeScript) to handle close event for all
subclasses and fire a close event explicitly if the close method is
called (before the internal dispose call).

Fix #2151

Change-Id: I403ea2c6e835560446576cd7fd5a0555640a0c8d