Prevent widget constructors from being marked as functional components
A widget factory generated via asFactory() may be passed a "parent"
factory as the second parameter to make it work as a selector. If the
given function was not actually a factory but a widget constructor
this would override the existing functionality of that constructor
as a selector, resulting in very confusing behavior.

Check that the parent factory is not a constructor for an object
that can be used as a JSX Element (and therefore as a selector).

Fix #2137

Change-Id: Ib0de0131436583f138adf8564914e30221771b18