Work around an unknown ArrayBuffer bug in iOS 14.5
In one Tabris.js application we ran into a case where an empty
ArrayBuffer was unexpectedly detached. That caused "Receiver is
detached" type error in `slice()` method.

Since we do not know for sure why and how that ArrayBuffer was happened
to be detached we implement a workaround to prevent this exception.

This issue was not reproducible in any lower version of iOS.

Change-Id: Id4216fdde4aa000ab3a4b8d996a588c0632d5297