Remove grunt task "verify_typings_fail"
This task took several minutes to execute, is difficult to understand,
and the tests are high maintenance.

It was introduced when we once had a regression where it become possible
to assign anything to a public method, killing the autocomplete effect
the typings are supposed to have. While it is valuable to prevent such
issues, it seems to be a very rare mistake and not worth the trade-offs.

Change-Id: I65897d415a9b4ca0442a468da21e3ded96233935