Make initial widget "data" object emit change events for all properties
This change makes the default "data" object that all widgets have emit
change events for all changes made to it. As a result the "data"
property doesn't have to be replaced entirely with a new object every
time the widget is supposed to update itself.

This is achieved via the new class "ObservableData", which uses a Proxy
to intercept all property changes so it can emit a matching change
event. Instances can be nested and will forward all change events to
the root object. As the name implies it also provides an Observable
instance, making it RxJS compatible.

To make this feature usable via declarative UI, make sure to not
replace the initial "data" object when set via an attribute if
possible.

The Widget and Observable classes are slightly refactored to share code
with ObservableData. EventObject and Listeners are tweaked to ensure no
unnecessary instances of "EventObject" are created when ObservableData
is listening to the change events of nested objects. Plain objects are
used instead (since they are faster) and only converted when accessed
via "originalEvent".

Documentation will follow in a later change together with making
ObservableData public.

Change-Id: I3861440ff8fdac026853ee728af1bb6918959601