Automatically apply "data" attribute after listeners
Normally all properties (including "data") within an attribute set (as
given to a JSX element or widget factory) are passed to the
constructor, then the listeners are registered. In case of "data" this
is very inconvenient since the data change listener may be used to
initialize the widget, but will not be invoked when data is already set.

There is now a special treatment for the "data" property to be set in
JsxProcessor after listeners are registered.

Revise functional component snippets to take advantage of this changed
behavior and re-arrange the code to showcase each use-case separately.

Change-Id: Iad3cb400e5b1393c5ce164adb798fbccb80fe5ec