Fix app crash when animating disposed widgets
The app crashes on Android when we animate the disposed widgets. The
Tabris Android platform now handles this case and returns a 'rejected'
value.

Abort animation and remove the `_dispose` listener when we receive the
'rejected' result from the native implementation.

Additionally, check the widget before starting animation and abort the
process when it is disposed.

Change-Id: I99cfcf9847b8c2d882f2de6389faeb7c3f82e74b