Make type declarations TypeScript 4.8 compatible
These changes are required to make the type declarations compile with
the TypeScript 4.8 compiler.

To do so the widget methods that return or accept a parent or sibling
widgets may no longer change that parent/sibling type through
inheritance. This is because it would create ambiguities when casting to
a supertype where the widget/parent type is different.

Child types are still known, but to prevent circular reference issues
the generic type parameter is widened to a new "AnyWidget" interface
that consists only of an unproblematic subset of the "Widget"
interface.

Further changes affect some selector APIs and generated widget Factory
interfaces.

These are minor breaking in changes, but unlikely to affect more than a
few lines even in larger applications, and the fix is to simply add a
cast or selector to the call to changed methods.

Many changes in the JSON files are due to the autoformatter.

Change-Id: I024828747e76af6b966f537aff599d664cfce624